lab done

+++ src/main/java/com/example/demo/repositories/PersonRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.repositories;
 import com.example.demo.models.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonRepository extends JpaRepository<Person, Long> {
+public interface PersonRepository extends JpaRepository<Person, Integer> {
 
 }
